requirements.txt
flask
yt-dlp
pydub
python-dotenv
gunicorn
ffmpeg-python




